Rename index specific classes
