--- app/models/manager_refresh/inventory_collection/index/storage/main.rb
@@ -1,15 +1,0 @@
-module ManagerRefresh
-  class InventoryCollection
-    module Index
-      class Storage::Main < ManagerRefresh::InventoryCollection::Index::Storage
-        def initialize(inventory_collection, attribute_names)
-          super
-        end
-
-        def find(index)
-          data_index[index]
-        end
-      end
-    end
-  end
-end
